update to react 14
