+++ packages/react-app/app/layout.tsx
@@ -1,0 +1,17 @@
+import '@/styles/globals.css';
+
+import { AppProvider } from '@/providers/AppProvider';
+
+export default function RootLayout({
+  children,
+}: Readonly<{
+  children: React.ReactNode;
+}>) {
+  return (
+    <html lang="en">
+      <body>
+        <AppProvider>{children}</AppProvider>
+      </body>
+    </html>
+  );
+}
