+++ packages/react-app/package.json
@@ -15,11 +15,12 @@
     "@tanstack/react-query": "^5.59.16",
     "@wagmi/connectors": "5.3.3",
     "encoding": "^0.1.13",
+    "eslint-config-next": "^15.0.2",
     "ethers": "^6.13.4",
-    "next": "^15.0.1",
+    "next": "^15.0.2",
     "pino-pretty": "^11.3.0",
-    "react": "18.3.1",
-    "react-dom": "18.3.1",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
     "viem": "2.21.34",
     "wagmi": "2.12.25"
   },
