--- packages/react-app/pages/_app.tsx
@@ -1,53 +1,0 @@
-import {
-  RainbowKitProvider,
-  connectorsForWallets,
-} from "@rainbow-me/rainbowkit";
-import "@rainbow-me/rainbowkit/styles.css";
-import { injectedWallet } from "@rainbow-me/rainbowkit/wallets";
-import type { AppProps } from "next/app";
-import { WagmiProvider, createConfig, http } from "wagmi";
-import { celo, celoAlfajores } from "wagmi/chains";
-import Layout from "../components/Layout";
-import "../styles/globals.css";
-
-import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
-
-const connectors = connectorsForWallets(
-  [
-    {
-      groupName: "Recommended",
-      wallets: [injectedWallet],
-    },
-  ],
-  {
-    appName: "Celo Composer",
-    projectId: process.env.WC_PROJECT_ID ?? "044601f65212332475a09bc14ceb3c34",
-  }
-);
-
-const config = createConfig({
-  connectors,
-  chains: [celo, celoAlfajores],
-  transports: {
-    [celo.id]: http(),
-    [celoAlfajores.id]: http(),
-  },
-});
-
-const queryClient = new QueryClient();
-
-function App({ Component, pageProps }: AppProps) {
-  return (
-    <WagmiProvider config={config}>
-      <QueryClientProvider client={queryClient}>
-        <RainbowKitProvider>
-          <Layout>
-            <Component {...pageProps} />
-          </Layout>
-        </RainbowKitProvider>
-      </QueryClientProvider>
-    </WagmiProvider>
-  );
-}
-
-export default App;
